requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
passlib
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
pydantic
pydantic-settings
fastapi-mail
email-validator
requests
python-dotenv